+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.wemo.roadassistant" >
-
+    package="com.wemo.roadassistant">
     <!--
          The ACCESS_COARSE/FINE_LOCATION permissions are not required to use
          Google Maps Android API v2, but you must specify either coarse or fine
@@ -21,8 +20,8 @@
         android:label="@string/app_name"
         android:roundIcon="@drawable/final_logo"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme" >
-
+        android:theme="@style/AppTheme">
+        <activity android:name=".FormActivity"></activity>
         <!--
              The API key for Google Maps-based APIs is defined as a string resource.
              (See the file "res/values/google_maps_api.xml").
@@ -37,9 +36,8 @@
 
         <activity
             android:name=".MapsActivity"
-            android:label="@string/title_activity_maps" >
-        </activity>
-        <activity android:name=".LoginRegistrationActivity" >
+            android:label="@string/title_activity_maps"></activity>
+        <activity android:name=".LoginRegistrationActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
