+++ app/src/main/res/layout/activity_main.xml
@@ -5,15 +5,16 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:layout_margin="20dp"
+    android:padding="20dp"
     android:gravity="center"
+    android:background="@drawable/backsacreen"
     tools:context=".LoginRegistrationActivity">
 
     <ImageView
         android:id="@+id/iv_avatar"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@mipmap/ic_launcher"/>
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+        android:src="@drawable/final_logo"/>
 
     <EditText
         android:id="@+id/et_email"
@@ -37,10 +38,12 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
         android:text="Login"
+        android:background="@color/colorAccent"
         android:textAllCaps="false"/>
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
         android:orientation="horizontal">
 
         <TextView
